// react-ui/types/generated/models/resource-type.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export type ResourceType = 'Sheep' | 'Wood' | 'Ore' | 'Wheat' | 'Brick' | 'GoldMine' | 'Desert' | 'Back' | 'None' | 'Sea' | 'Coin' | 'Cloth' | 'Paper' | 'Politics' | 'Trade' | 'Science' | 'AnyDevCard' | 'VictoryPoint' | 'Invasion' | 'Robber';

export const ResourceType = {
    Sheep: 'Sheep',
    Wood: 'Wood',
    Ore: 'Ore',
    Wheat: 'Wheat',
    Brick: 'Brick',
    GoldMine: 'GoldMine',
    Desert: 'Desert',
    Back: 'Back',
    None: 'None',
    Sea: 'Sea',
    Coin: 'Coin',
    Cloth: 'Cloth',
    Paper: 'Paper',
    Politics: 'Politics',
    Trade: 'Trade',
    Science: 'Science',
    AnyDevCard: 'AnyDevCard',
    VictoryPoint: 'VictoryPoint',
    Invasion: 'Invasion',
    Robber: 'Robber',
} as const;